More fixes
1. add setter for clipBehavior to prevent Clip.none
2. move ClipContext to painting layer
