+++ packages/flutter/lib/painting.dart
@@ -28,6 +28,7 @@ export 'src/painting/box_decoration.dart';
 export 'src/painting/box_fit.dart';
 export 'src/painting/box_shadow.dart';
 export 'src/painting/circle_border.dart';
+export 'src/painting/clip.dart';
 export 'src/painting/colors.dart';
 export 'src/painting/debug.dart';
 export 'src/painting/decoration.dart';
