+++ packages/flutter/lib/src/rendering/object.dart
@@ -45,61 +45,6 @@ class ParentData {
 /// Used by many of the methods of [PaintingContext].
 typedef void PaintingContextCallback(PaintingContext context, Offset offset);
 
-/// Clip utilities used by [PaintingContext] and [TestRecordingPaintingContext].
-abstract class ClipContext {
-  /// The canvas on which to paint.
-  Canvas get canvas;
-
-  // TODO(liyuqian): change argument order
-  void _clipAndPaint(void canvasClipCall(bool doAntiAlias), Clip clipBehavior, Rect bounds, void painter()) {
-    assert(canvasClipCall != null);
-    canvas.save();
-    switch (clipBehavior) {
-      case Clip.none:
-        break;
-      case Clip.hardEdge:
-        canvasClipCall(false);
-        break;
-      case Clip.antiAlias:
-        canvasClipCall(true);
-        break;
-      case Clip.antiAliasWithSaveLayer:
-        canvasClipCall(true);
-        canvas.saveLayer(bounds, new Paint());
-        break;
-    }
-    painter();
-    if (clipBehavior == Clip.antiAliasWithSaveLayer) {
-      canvas.restore();
-    }
-    canvas.restore();
-  }
-
-  /// Clip [canvas] with [Path] according to [Clip] and then paint. [canvas] is
-  /// restored to the pre-clip status afterwards.
-  ///
-  /// `bounds` is the saveLayer bounds used for [Clip.antiAliasWithSaveLayer].
-  void clipPathAndPaint(Path path, Clip clipBehavior, Rect bounds, void painter()) {
-    _clipAndPaint((bool doAntiAias) => canvas.clipPath(path, doAntiAlias: doAntiAias), clipBehavior, bounds, painter);
-  }
-
-  /// Clip [canvas] with [Path] according to [RRect] and then paint. [canvas] is
-  /// restored to the pre-clip status afterwards.
-  ///
-  /// `bounds` is the saveLayer bounds used for [Clip.antiAliasWithSaveLayer].
-  void clipRRectAndPaint(RRect rrect, Clip clipBehavior, Rect bounds, void painter()) {
-    _clipAndPaint((bool doAntiAias) => canvas.clipRRect(rrect, doAntiAlias: doAntiAias), clipBehavior, bounds, painter);
-  }
-
-  /// Clip [canvas] with [Path] according to [Rect] and then paint. [canvas] is
-  /// restored to the pre-clip status afterwards.
-  ///
-  /// `bounds` is the saveLayer bounds used for [Clip.antiAliasWithSaveLayer].
-  void clipRectAndPaint(Rect rect, Clip clipBehavior, Rect bounds, void painter()) {
-    _clipAndPaint((bool doAntiAias) => canvas.clipRect(rect, doAntiAlias: doAntiAias), clipBehavior, bounds, painter);
-  }
-}
-
 /// A place to paint.
 ///
 /// Rather than holding a canvas directly, [RenderObject]s paint using a painting
